crewAI/search/requirements.txt
requests
python-dotenv
crewai
flask